+++ pkg/target/client.go
@@ -4,6 +4,7 @@ import (
 	"strings"
 
 	"github.com/kyverno/policy-reporter/pkg/report"
+	"github.com/minio/pkg/wildcard"
 )
 
 // Client for a provided Target
@@ -22,34 +23,121 @@ type Client interface {
 	Sources() []string
 }
 
+type Rules struct {
+	Exclude []string
+	Include []string
+}
+
+type Filter struct {
+	Namespace       Rules
+	Priority        Rules
+	Policy          Rules
+	MinimumPriority string
+	Sources         []string
+}
+
+func (f *Filter) Validate(result *report.Result) bool {
+	if len(f.Sources) > 0 && !contains(result.Source, f.Sources) {
+		return false
+	}
+
+	if result.Priority < report.NewPriority(f.MinimumPriority) {
+		return false
+	}
+
+	if !f.validateNamespaceRules(result) {
+		return false
+	}
+
+	if !f.validatePolicyRules(result) {
+		return false
+	}
+
+	if !f.validatePriorityRules(result) {
+		return false
+	}
+
+	return true
+}
+
+func (f *Filter) validateNamespaceRules(result *report.Result) bool {
+	if result.Resource != nil && len(f.Namespace.Include) > 0 {
+		for _, ns := range f.Namespace.Include {
+			if wildcard.Match(ns, result.Resource.Namespace) {
+				return true
+			}
+		}
+
+		return false
+	} else if result.Resource != nil && len(f.Namespace.Exclude) > 0 {
+		for _, ns := range f.Namespace.Exclude {
+			if wildcard.Match(ns, result.Resource.Namespace) {
+				return false
+			}
+		}
+	}
+
+	return true
+}
+
+func (f *Filter) validatePolicyRules(result *report.Result) bool {
+	if len(f.Policy.Include) > 0 {
+		for _, ns := range f.Policy.Include {
+			if wildcard.Match(ns, result.Policy) {
+				return true
+			}
+		}
+
+		return false
+	} else if len(f.Policy.Exclude) > 0 {
+		for _, ns := range f.Policy.Exclude {
+			if wildcard.Match(ns, result.Policy) {
+				return false
+			}
+		}
+	}
+
+	return true
+}
+
+func (f *Filter) validatePriorityRules(result *report.Result) bool {
+	if len(f.Priority.Include) > 0 {
+		return contains(result.Priority.String(), f.Priority.Include)
+	} else if len(f.Priority.Exclude) > 0 && contains(result.Priority.String(), f.Priority.Exclude) {
+		return false
+	}
+
+	return true
+}
+
 type BaseClient struct {
-	minimumPriority       string
-	sources               []string
+	name                  string
 	skipExistingOnStartup bool
+	filter                *Filter
+}
+
+func (c *BaseClient) Name() string {
+	return c.name
 }
 
 func (c *BaseClient) MinimumPriority() string {
-	return c.minimumPriority
+	return c.filter.MinimumPriority
 }
 
 func (c *BaseClient) Sources() []string {
-	return c.sources
+	return c.filter.Sources
+}
+
+func (c *BaseClient) Validate(result *report.Result) bool {
+	return c.filter.Validate(result)
 }
 
 func (c *BaseClient) SkipExistingOnStartup() bool {
 	return c.skipExistingOnStartup
 }
 
-func (c *BaseClient) Validate(result *report.Result) bool {
-	if result.Priority < report.NewPriority(c.minimumPriority) {
-		return false
-	}
-
-	if len(c.sources) > 0 && !contains(result.Source, c.sources) {
-		return false
-	}
-
-	return true
+func NewBaseClient(name string, skipExistingOnStartup bool, filter *Filter) BaseClient {
+	return BaseClient{name, skipExistingOnStartup, filter}
 }
 
 func contains(source string, sources []string) bool {
@@ -60,8 +148,4 @@ func contains(source string, sources []string) bool {
 	}
 
 	return false
-}
-
-func NewBaseClient(minimumPriority string, sources []string, skipExistingOnStartup bool) BaseClient {
-	return BaseClient{minimumPriority, sources, skipExistingOnStartup}
 }