+++ pkg/target/discord/discord_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/kyverno/policy-reporter/pkg/report"
+	"github.com/kyverno/policy-reporter/pkg/target"
 	"github.com/kyverno/policy-reporter/pkg/target/discord"
 )
 
@@ -67,7 +68,7 @@ func Test_LokiTarget(t *testing.T) {
 			}
 		}
 
-		client := discord.NewClient("http://hook.discord:80", "", []string{}, false, testClient{callback, 200})
+		client := discord.NewClient("Discord", "http://hook.discord:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(completeResult)
 	})
 
@@ -86,11 +87,11 @@ func Test_LokiTarget(t *testing.T) {
 			}
 		}
 
-		client := discord.NewClient("http://hook.discord:80", "", []string{}, false, testClient{callback, 200})
+		client := discord.NewClient("Discord", "http://hook.discord:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(minimalResult)
 	})
 	t.Run("Name", func(t *testing.T) {
-		client := discord.NewClient("http://localhost:9200", "", []string{}, true, testClient{})
+		client := discord.NewClient("Discord", "http://localhost:9200", true, &target.Filter{}, testClient{})
 
 		if client.Name() != "Discord" {
 			t.Errorf("Unexpected Name %s", client.Name())