+++ pkg/target/loki/loki_test.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/kyverno/policy-reporter/pkg/report"
+	"github.com/kyverno/policy-reporter/pkg/target"
 	"github.com/kyverno/policy-reporter/pkg/target/loki"
 )
 
@@ -117,7 +118,7 @@ func Test_LokiTarget(t *testing.T) {
 			}
 		}
 
-		loki := loki.NewClient("http://localhost:3100", "", []string{}, false, map[string]string{"custom": "label"}, testClient{callback, 200})
+		loki := loki.NewClient("Loki", "http://localhost:3100", false, &target.Filter{}, map[string]string{"custom": "label"}, testClient{callback, 200})
 		loki.Send(completeResult)
 	})
 
@@ -175,11 +176,11 @@ func Test_LokiTarget(t *testing.T) {
 			}
 		}
 
-		loki := loki.NewClient("http://localhost:3100", "", []string{}, false, make(map[string]string), testClient{callback, 200})
+		loki := loki.NewClient("Loki", "http://localhost:3100", false, &target.Filter{}, make(map[string]string), testClient{callback, 200})
 		loki.Send(minimalResult)
 	})
 	t.Run("Name", func(t *testing.T) {
-		client := loki.NewClient("http://localhost:9200", "", []string{}, true, make(map[string]string), testClient{})
+		client := loki.NewClient("Loki", "http://localhost:9200", true, &target.Filter{}, make(map[string]string), testClient{})
 
 		if client.Name() != "Loki" {
 			t.Errorf("Unexpected Name %s", client.Name())